Initial commit
Initial work on an NES emulator implemented in C. This is still a mess,
but contains already some structure in terms of how unit tests will be
laid out (using Check), as well as how the UI will work (using imgui).

For now, the UI code is still C++, this will be rewritten at a later
stage to use cimgui.

The only 'useful' tool right now is the rom dump tool, that is able to
dump some iNes ROM information to standard output.