Improve scrolling; add status line
The textual assembly is now only printed once to one big notcurses
plane, which is subsequently moved around to simulate scrolling. This is
simpler (and faster?) than reprinting all of the assembly every time.

Adds a simple status line with a very basic help screen.

When manually inputting an address to jump to, backspace actually works.