Initial work on CPU emulation
Lots of stuff going on; some instructions have been implemented.

The debugger learned two commands, 'next instruction' ('n') and run
('r'), which will either execute one instruction and wait again for
input, or will just start code execution and wait for the debugger to
eventually exit because of an unknown opcode encountered ;)

The debugger can now write out a log file that resembles the format of
the famous Nintendulator golden reference log for `nestest.rom`. It
lacks the PPU part for now, since that part will be implemented later.

This disables notcurses not printing out a banner for now because of a
bug that I encountered, see:

https://github.com/dankamongmen/notcurses/issues/1314

Small fixes in the CPU power on and reset states; the cycle count was
incorrectly updated.