dbg: major improvements
Scrolling in the debugger is now done in the same way as it was done in
the initial version; for every change in memory location, the data that
needs to be printed is recalculated again instead of loaded from a
cache. For now, this is more than fast enough, and it prevents issues
where the program writes to memory and the debugger does not see it,
because it cached the whole assembly view...ooops.

Furthermore, the debugger now only interprets a subset of memory as
assembly; that part covered by the loaded ROM (for now). In the future,
surely this needs to be adapted in case of self-modifying code etc., but
for now this works and prevents issues when programs start storing data
in zero page memory for example.

This adds emulation for a whole bunch of branch instructions to the CPU.

Add a new command 'f', which will bring the focus of the assembly window
back to the program counter.

Fixes a small formatting issue in the Nintendulator log.