dbg: add support for more CPU instructions
Implement a whole bunch of additional CPU instructions. The first 1200
cycles or so of the NES test ROM are now correctly emulated.

Fix a couple of issues with updating the negative and zero flags,
simplify the code for this as well.

The add and subtract with carry instructions are known to be broken.