cpu: implement additional instructions
Implement STX, and implement absolute addressing mode versions of LDA,
LDX, and STX. Hack together a way to keep the Nintendulator mode log
compatible with the actual Nintendulator output. It seems the value at
a memory location for an instruction that uses absolute addressing mode
is only printed for load/save instructions (?). Need to verify this.