Implement 'break at cycle'
Allow the user to specify a certain cycle number where the debugger
should stop execution, and query the user for input again. Useful to
have now that we reach higher cycle numbers in the NES test ROM.

Furthermore; found that the notcurses issue reported previously is still
present when querying the user for input :( Still need to report this.