dbg: add breakpoint functionality
Add a way to define breakpoints in `dbg`. Using space bar in the
assembly view will now toggle breakpoints. More than one breakpoint may
be set at the same time. The breakpoints pane shows an overview of all
breakpoints that are set. It is possible to shift focus from the
assembly pane to the breakpoints pane using `<tab>`. In the breakpoints
pane, it is possible to remove breakpoints using `d`. `c` clears all
breakpoints.

This also moves a bunch of UI code from `dbg/main.c` to different pane
implementations (breakpoints, status, and CPU pane). Better separation
of concerns and allows for easier generalization of code in the future.