Remove a lot of unneeded endianness conversions
The code now assumes that the target machine's endianness is the same as
the 6502 endianness (little), which happens to be the case for x86, but
we should get rid of this limitation! In any case, the previous code was
bad, and this requirement can be satisfied in a more elegant way.